server/requirements.txt
grpcio-tools
googleapis-common-protos